+++ lib/sendgrid4r/rest/api.rb
@@ -29,13 +29,13 @@ require 'sendgrid4r/rest/contacts/reserved_fields'
 require 'sendgrid4r/rest/contacts/segments'
 require 'sendgrid4r/rest/campaigns/campaigns'
 require 'sendgrid4r/rest/api_keys/api_keys'
-require 'sendgrid4r/rest/subusers/subusers'
+require 'sendgrid4r/rest/subusers'
 require 'sendgrid4r/rest/email_activity/email_activity'
 require 'sendgrid4r/rest/whitelabel/domains'
 require 'sendgrid4r/rest/whitelabel/ips'
 require 'sendgrid4r/rest/whitelabel/links'
-require 'sendgrid4r/rest/users/users'
-require 'sendgrid4r/rest/bounces/bounces'
+require 'sendgrid4r/rest/users'
+require 'sendgrid4r/rest/bounces'
 
 module SendGrid4r
   module REST
