+++ lib/sendgrid4r/rest/users.rb
