+++ spec/rest/bounces_spec.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-require File.dirname(__FILE__) + '/../../spec_helper'
+require File.dirname(__FILE__) + '/../spec_helper'
 
 describe SendGrid4r::REST::Bounces do
   describe 'integration test', :it do
